requirements.txt
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
gspread>=6.0.0
google-auth>=2.27.0
python-multipart>=0.0.6
twilio>=8.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
